+++ ui/src/components/DashboardServices.js
@@ -3,7 +3,8 @@ import React from 'react';
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
 import { useIntl } from 'react-intl';
-import { Tooltip, StatusText } from '@scality/core-ui';
+import Tooltip from '@scality/core-ui/dist/components/tooltip/Tooltip.component';
+import { StatusText } from '@scality/core-ui/dist/components/text/Text.component';
 import {
   spacing,
   fontSize,