+++ ui/src/containers/App.js
@@ -1,20 +1,23 @@
 //@flow
 import React from 'react';
+import { MetricsTimeSpanProvider } from '@scality/core-ui/dist/next';
 import Layout from './Layout';
 
 import AlertProvider from './AlertProvider';
 import ConfigProvider from './ConfigProvider';
-import { MetricsTimeSpanProvider } from '../hooks';
 import FederatedIntlProvider from './IntlProvider';
+import StartTimeProvider from './StartTimeProvider';
 
 const App = () => {
   return (
     <FederatedIntlProvider>
       <MetricsTimeSpanProvider>
         <ConfigProvider>
-          <AlertProvider>
-            <Layout />
-          </AlertProvider>
+          <StartTimeProvider>
+            <AlertProvider>
+              <Layout />
+            </AlertProvider>
+          </StartTimeProvider>
         </ConfigProvider>
       </MetricsTimeSpanProvider>
     </FederatedIntlProvider>