+++ ui/src/containers/NodePageRSP.js
@@ -8,8 +8,6 @@ import { getPodsListData } from '../services/PodUtils';
 import { useURLQuery, useRefreshEffect } from '../services/utils';
 import {
   updateNodeStatsFetchArgumentAction,
-  refreshNodeStatsAction,
-  stopRefreshNodeStatsAction,
   fetchNodeUNameInfoAction,
 } from '../ducks/app/monitoring';
 import {
@@ -35,7 +33,6 @@ import {
 } from '../constants';
 import { useAlerts } from './AlertProvider';
 import { useIntl } from 'react-intl';
-import { useTypedSelector } from '../hooks';
 
 // <NodePageRSP> fetches the data for all the tabs given the current selected Node
 // handles the refresh for the metrics tab
@@ -70,7 +67,6 @@ const NodePageRSP = (props) => {
 
   const showAvg = queryShowAvg === 'true' ? true : nodeMetricsShowAvg;
 
-  useRefreshEffect(refreshNodeStatsAction, stopRefreshNodeStatsAction);
   useRefreshEffect(refreshVolumesAction, stopRefreshVolumesAction);
 
   // Retrieve the podlist data
@@ -78,18 +74,14 @@ const NodePageRSP = (props) => {
   const podsListData = getPodsListData(name, pods);
   const nodes = useSelector((state) => state.app.nodes.list);
   const volumes = useSelector((state) => state.app.volumes.list);
-  const nodeStats = useTypedSelector(
-    (state) => state.app.monitoring.nodeStats.metrics,
-  );
-  const avgStats = useTypedSelector(
-    (state) => state.app.monitoring.nodeStats.metricsAvg,
-  );
 
   const nodesIPsInfo = useSelector((state) => state.app.nodes.IPsInfo);
   const instanceIP =
     nodes?.find((node) => node.name === name)?.internalIP ?? '';
-  const controlPlaneInterface = nodesIPsInfo[name]?.controlPlane?.interface;
-  const workloadPlaneInterface = nodesIPsInfo[name]?.workloadPlane?.interface;
+  const controlPlaneInterface =
+    nodesIPsInfo[name]?.controlPlane?.interface ?? '';
+  const workloadPlaneInterface =
+    nodesIPsInfo[name]?.workloadPlane?.interface ?? '';
   const currentNode = nodeTableData?.find((node) => node.name.name === name);
 
   useEffect(() => {
@@ -159,9 +151,10 @@ const NodePageRSP = (props) => {
       >
         <NodePageMetricsTab
           nodeName={name}
-          nodeStats={nodeStats}
-          avgStats={avgStats}
           instanceIP={instanceIP}
+          controlPlaneInterface={controlPlaneInterface}
+          workloadPlaneInterface={workloadPlaneInterface}
+          nodesIPsInfo={nodesIPsInfo}
         />
       </Tabs.Tab>
       <Tabs.Tab